+++ apps/decodex-app/Sources/DecodexApp/AccountUsageMeterView.swift
@@ -1,0 +1,190 @@
+import Foundation
+import SwiftUI
+
+struct AccountUsageMeterView: View {
+	let label: String
+	let remainingPercent: Int?
+	let resetAtUnixEpoch: Int?
+	let dailyAveragePercent: Double?
+	let tone: AccountTone
+	let currentTime: Date
+	@Environment(\.colorScheme) private var colorScheme
+
+	var body: some View {
+		VStack(alignment: .leading, spacing: 3) {
+			HStack(spacing: 5) {
+				Text(label)
+					.font(PanelFont.usageLabel)
+					.frame(width: 28, alignment: .leading)
+					.foregroundStyle(PanelPalette.secondaryText(colorScheme))
+
+				Text(remainingText)
+					.font(PanelFont.usageValue)
+					.frame(width: 62, alignment: .leading)
+					.foregroundStyle(valueColor)
+					.monospacedDigit()
+
+				if let dailyAverageText {
+					HStack(alignment: .firstTextBaseline, spacing: 3) {
+						Text("avg")
+							.font(PanelFont.usageLabel)
+							.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(0.82))
+							.lineLimit(1)
+
+						Text(dailyAverageText)
+							.font(PanelFont.usageValue)
+							.foregroundStyle(PanelPalette.secondaryText(colorScheme))
+							.monospacedDigit()
+							.lineLimit(1)
+							.minimumScaleFactor(0.78)
+					}
+					.layoutPriority(1)
+				}
+
+				Spacer(minLength: 2)
+
+				Text(resetDisplay.short)
+					.font(PanelFont.usageValue)
+					.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(colorScheme == .dark ? 0.82 : 0.9))
+					.monospacedDigit()
+					.lineLimit(1)
+
+				if resetDisplay.date.isEmpty == false {
+					Text(resetDisplay.date)
+						.font(PanelFont.tertiary)
+						.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(colorScheme == .dark ? 0.68 : 0.78))
+						.lineLimit(1)
+						.truncationMode(.middle)
+					}
+			}
+			.frame(height: 14)
+
+			GeometryReader { proxy in
+				ZStack(alignment: .leading) {
+					let width = fillWidth(in: proxy.size.width)
+
+					Capsule()
+						.fill(trackColor)
+						.overlay {
+							Capsule()
+								.fill(trackInsetStyle)
+								.padding(.vertical, 0.8)
+								.allowsHitTesting(false)
+						}
+					Capsule()
+						.fill(fillStyle)
+						.frame(width: width)
+						.clipShape(Capsule())
+						.animation(PanelMotion.state, value: remainingPercent)
+						.shadow(
+							color: color.opacity(colorScheme == .dark ? 0.09 : 0.07),
+							radius: colorScheme == .dark ? 1.2 : 1,
+							x: 0,
+							y: 0
+						)
+					Capsule()
+						.strokeBorder(trackEdgeColor, lineWidth: 0.24)
+						.allowsHitTesting(false)
+				}
+			}
+			.frame(height: 3.2)
+		}
+		.lineLimit(1)
+		.frame(height: 22)
+		.frame(maxWidth: .infinity, alignment: .leading)
+		.accessibilityLabel(accessibilityText)
+	}
+
+	private var remainingText: String {
+		guard let remainingPercent else {
+			return "-"
+		}
+
+		return "\(remainingPercent)% left"
+	}
+
+	private var dailyAverageText: String? {
+		guard let dailyAveragePercent else {
+			return nil
+		}
+		let formatted = formatDailyUsageRate(dailyAveragePercent)
+
+		return formatted == "-" ? nil : formatted
+	}
+
+	private var accessibilityText: String {
+		let average = dailyAverageText.map { ", daily average \($0)" } ?? ""
+		return "\(label) remaining \(remainingText)\(average), \(resetDisplay.accessibility)"
+	}
+
+	private var progress: CGFloat {
+		guard let remainingPercent else {
+			return 0
+		}
+
+		return CGFloat(max(0, min(100, remainingPercent))) / 100
+	}
+
+	private func fillWidth(in width: CGFloat) -> CGFloat {
+		guard remainingPercent != nil else {
+			return 0
+		}
+
+		return max(4, width * progress)
+	}
+
+	private var color: Color {
+		switch tone {
+		case .codexActive: return PanelPalette.codexAccent(colorScheme)
+		case .ready: return PanelPalette.capacityAccent(colorScheme)
+		case .selected: return PanelPalette.routeAccent(colorScheme)
+		case .warning: return PanelPalette.warning(colorScheme)
+		case .danger: return PanelPalette.destructive(colorScheme)
+		case .neutral: return PanelPalette.secondaryText(colorScheme)
+		}
+	}
+
+	private var valueColor: Color {
+		switch tone {
+		case .warning, .danger:
+			return color.opacity(colorScheme == .dark ? 0.95 : 0.78)
+		default:
+			return PanelPalette.primaryText(colorScheme).opacity(colorScheme == .dark ? 0.9 : 0.84)
+		}
+	}
+
+	private var resetDisplay: UsageResetDisplay {
+		UsageResetDisplay.make(resetAtUnixEpoch: resetAtUnixEpoch, now: currentTime)
+	}
+
+	private var trackColor: Color {
+		PanelPalette.progressTrack(colorScheme)
+	}
+
+	private var trackEdgeColor: Color {
+		PanelPalette.progressEdge(colorScheme)
+	}
+
+	private var fillStyle: LinearGradient {
+		LinearGradient(
+			colors: [
+				color.opacity(colorScheme == .dark ? 0.78 : 0.68),
+				color.opacity(colorScheme == .dark ? 0.62 : 0.52),
+			],
+			startPoint: .leading,
+			endPoint: .trailing
+		)
+	}
+
+	private var trackInsetStyle: LinearGradient {
+		LinearGradient(
+			colors: [
+				Color.white.opacity(colorScheme == .dark ? 0.022 : 0.05),
+				Color.white.opacity(0),
+				Color.black.opacity(colorScheme == .dark ? 0.035 : 0.018),
+			],
+			startPoint: .top,
+			endPoint: .bottom
+		)
+	}
+}
