+++ apps/decodex-app/Sources/DecodexApp/NoticeView.swift
@@ -1,0 +1,22 @@
+import SwiftUI
+
+struct NoticeView: View {
+	let text: String
+	@Environment(\.colorScheme) private var colorScheme
+
+	var body: some View {
+		HStack(alignment: .top, spacing: 7) {
+			Image(systemName: "exclamationmark.triangle")
+				.foregroundStyle(PanelPalette.warning(colorScheme))
+			Text(text)
+				.font(PanelFont.notice)
+				.foregroundStyle(PanelPalette.secondaryText(colorScheme))
+				.fixedSize(horizontal: false, vertical: true)
+		}
+		.padding(8)
+		.modernGlassSurface(
+			cornerRadius: 9,
+			depth: .section
+		)
+	}
+}
